+++ spec/unit/coinbase/wallet_spec.rb
@@ -4,29 +4,36 @@ describe Coinbase::Wallet do
   subject(:wallet) { described_class.new(model) }
 
   let(:wallet_id) { SecureRandom.uuid }
-  let(:network) { :base_sepolia }
-  let(:network_id) { Coinbase.normalize_network(network) }
+  let(:network_id) { :base_sepolia }
+  let(:normalized_network_id) { Coinbase.normalize_network(network_id) }
   let(:seed) { '000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f' }
-  let(:model) { build(:wallet_model, network, :without_default_address, id: wallet_id) }
-  let(:model_with_default_address) { build(:wallet_model, network, id: wallet_id, seed: seed) }
+  let(:model) { build(:wallet_model, network_id, :without_default_address, id: wallet_id) }
+  let(:model_with_default_address) { build(:wallet_model, network_id, id: wallet_id, seed: seed) }
   let(:model_with_seed_pending) do
-    build(:wallet_model, network, :server_signer_pending, id: wallet_id, seed: seed)
+    build(:wallet_model, network_id, :server_signer_pending, id: wallet_id, seed: seed)
   end
   let(:model_with_seed_active) do
-    build(:wallet_model, network, :server_signer_active, id: wallet_id, seed: seed)
+    build(:wallet_model, network_id, :server_signer_active, id: wallet_id, seed: seed)
   end
   let(:first_address_model) do
-    build(:address_model, network, :with_seed, seed: seed, wallet_id: wallet_id, index: 0)
+    build(:address_model, network_id, :with_seed, seed: seed, wallet_id: wallet_id, index: 0)
   end
   let(:second_address_model) do
-    build(:address_model, network, :with_seed, seed: seed, wallet_id: wallet_id, index: 1)
+    build(:address_model, network_id, :with_seed, seed: seed, wallet_id: wallet_id, index: 1)
   end
   let(:wallets_api) { instance_double(Coinbase::Client::WalletsApi) }
   let(:addresses_api) { instance_double(Coinbase::Client::AddressesApi) }
   let(:transfers_api) { instance_double(Coinbase::Client::TransfersApi) }
   let(:use_server_signer) { false }
+  let(:default_network) { build(:network, :base_sepolia) }
+  let(:network) { default_network }
   let(:configuration) do
-    instance_double(Coinbase::Configuration, use_server_signer: use_server_signer, api_client: nil)
+    instance_double(
+      Coinbase::Configuration,
+      use_server_signer: use_server_signer,
+      api_client: nil,
+      default_network: default_network
+    )
   end
 
   before do
@@ -34,6 +41,11 @@ describe Coinbase::Wallet do
     allow(Coinbase::Client::WalletsApi).to receive(:new).and_return(wallets_api)
 
     allow(Coinbase).to receive(:configuration).and_return(configuration)
+
+    allow(Coinbase::Network)
+      .to receive(:from_id)
+      .with(satisfy { |id| id == network_id || id == normalized_network_id })
+      .and_return(network)
   end
 
   describe '.list' do
@@ -44,7 +56,7 @@ describe Coinbase::Wallet do
     let(:resource_list_klass) { Coinbase::Client::WalletList }
     let(:item_klass) { described_class }
     let(:item_initialize_args) { { seed: '' } }
-    let(:create_model) { ->(id) { build(:wallet_model, network, :without_default_address, id: id) } }
+    let(:create_model) { ->(id) { build(:wallet_model, network_id, :without_default_address, id: id) } }
 
     it_behaves_like 'it is a paginated enumerator', :wallets
   end
@@ -79,7 +91,7 @@ describe Coinbase::Wallet do
     subject(:imported_wallet) { described_class.import(exported_data) }
 
     let(:wallet_id) { SecureRandom.uuid }
-    let(:wallet_model) { build(:wallet_model, network, id: wallet_id) }
+    let(:wallet_model) { build(:wallet_model, network_id, id: wallet_id) }
     let(:exported_data) { Coinbase::Wallet::Data.new(wallet_id: wallet_id, seed: seed) }
 
     before do
@@ -126,10 +138,10 @@ describe Coinbase::Wallet do
 
     let(:wallet_id) { SecureRandom.uuid }
     let(:create_wallet_request) do
-      { wallet: { network_id: network_id, use_server_signer: use_server_signer } }
+      { wallet: { network_id: normalized_network_id, use_server_signer: use_server_signer } }
     end
     let(:request) { { create_wallet_request: create_wallet_request } }
-    let(:wallet_model) { build(:wallet_model, network, id: wallet_id) }
+    let(:wallet_model) { build(:wallet_model, network_id, id: wallet_id) }
 
     before do
       allow(wallets_api).to receive(:create_wallet).with(request).and_return(wallet_model)
@@ -178,13 +190,15 @@ describe Coinbase::Wallet do
       end
     end
 
-    context 'when setting the network ID explicitly' do
+    context 'when setting the network explicitly' do
       subject(:created_wallet) do
-        described_class.create(network_id: network)
+        described_class.create(network: network)
       end
 
-      let(:network) { :base_mainnet }
+      let(:network_id) { :ethereum_mainnet }
       let(:use_server_signer) { false }
+      let(:network) { build(:network, network_id) }
+      let(:normalized_network_id) { 'ethereum-mainnet' }
 
       before do
         allow(addresses_api)
@@ -210,12 +224,18 @@ describe Coinbase::Wallet do
         expect(created_wallet).to be_a(described_class)
       end
 
-      it 'sets the specified network ID' do
-        expect(created_wallet.network_id).to eq(:base_mainnet)
+      it 'sets the specified network' do
+        expect(created_wallet.network).to eq(network)
       end
 
       context 'when using a network symbol' do
+        subject(:created_wallet) do
+          described_class.create(network: network_id)
+        end
+
+        let(:normalized_network_id) { 'base-mainnet' }
         let(:network_id) { :base_mainnet }
+        let(:network) { build(:network, network_id) }
         let(:create_wallet_request) do
           { wallet: { network_id: 'base-mainnet', use_server_signer: use_server_signer } }
         end
@@ -224,8 +244,8 @@ describe Coinbase::Wallet do
           expect(created_wallet).to be_a(described_class)
         end
 
-        it 'sets the specified network ID' do
-          expect(created_wallet.network_id).to eq(:base_mainnet)
+        it 'sets the specified network' do
+          expect(created_wallet.network).to eq(network)
         end
       end
     end
@@ -262,7 +282,7 @@ describe Coinbase::Wallet do
       end
 
       it 'sets the default network ID' do
-        expect(created_wallet.network_id).to eq(:base_sepolia)
+        expect(created_wallet.network).to eq(network)
       end
     end
 
@@ -395,8 +415,8 @@ describe Coinbase::Wallet do
   end
 
   describe '#network_id' do
-    it 'returns the Network ID' do
-      expect(wallet.network_id).to eq(:base_sepolia)
+    it 'returns the Network' do
+      expect(wallet.network).to eq(network)
     end
   end
 
@@ -779,8 +799,8 @@ describe Coinbase::Wallet do
     let(:response) do
       Coinbase::Client::AddressBalanceList.new(
         data: [
-          build(:balance_model, network, amount: '1000000000000000000'),
-          build(:balance_model, network, :usdc, amount: '5000000')
+          build(:balance_model, network_id, amount: '1000000000000000000'),
+          build(:balance_model, network_id, :usdc, amount: '5000000')
         ]
       )
     end
@@ -797,7 +817,7 @@ describe Coinbase::Wallet do
   describe '#balance' do
     let(:eth_asset) { build(:asset_model) }
     let(:amount) { 5_000_000_000_000_000_000 }
-    let(:response) { build(:balance_model, network, amount: amount) }
+    let(:response) { build(:balance_model, network_id, amount: amount) }
 
     before do
       allow(wallets_api).to receive(:get_wallet_balance).with(wallet_id, 'eth').and_return(response)
@@ -900,7 +920,7 @@ describe Coinbase::Wallet do
     let(:asset_id) { :eth }
     let(:other_wallet_id) { SecureRandom.uuid }
     let(:other_wallet_model) do
-      build(:wallet_model, network, id: other_wallet_id, default_address: second_address_model)
+      build(:wallet_model, network_id, id: other_wallet_id, default_address: second_address_model)
     end
     let(:other_wallet) { described_class.new(other_wallet_model, seed: '') }
     let(:destination) { other_wallet }
@@ -1223,7 +1243,7 @@ describe Coinbase::Wallet do
     it 'loads the encrypted seed from file with multiple seeds' do
       seed_wallet.save_seed!(file_path, encrypt: true)
 
-      other_model = Coinbase::Client::Wallet.new(id: SecureRandom.uuid, network_id: network_id)
+      other_model = Coinbase::Client::Wallet.new(id: SecureRandom.uuid, network_id: normalized_network_id)
       other_wallet = described_class.new(other_model)
       other_wallet.save_seed!(file_path, encrypt: true)
 
@@ -1270,7 +1290,7 @@ describe Coinbase::Wallet do
 
   describe '#inspect' do
     it 'includes wallet details' do
-      expect(wallet.inspect).to include(wallet_id, Coinbase.to_sym(network_id).to_s)
+      expect(wallet.inspect).to include(wallet_id, network_id.to_s)
     end
 
     it 'returns the same value as to_s' do